Tests passing with new base_inversion + mods to components.  Working through bugs running gravity inversion now.
